refactor: make the pixbuf upload reusable
Turn the renderer-private upload_background_image into
create_image_from_pixbuf in femtovg_area, so that tools drawing their own
pixbufs can upload them to the canvas as well
