+++ src/femtovg_area/mod.rs
@@ -2,7 +2,13 @@ mod imp;
 
 use std::{cell::RefCell, rc::Rc, sync::OnceLock};
 
-use femtovg::FontId;
+use anyhow::Result;
+use femtovg::{
+    Canvas, FontId, ImageFlags, ImageId, ImageSource, PixelFormat,
+    imgref::Img,
+    renderer::OpenGl,
+    rgb::{RGB, RGBA},
+};
 use gtk::glib;
 use relm4::gtk::gdk_pixbuf::{Pixbuf, glib::subclass::types::ObjectSubclassIsExt};
 use relm4::{
@@ -25,6 +31,73 @@ pub fn set_font_stack(fonts: Vec<FontId>) {
 
 pub fn font_stack() -> &'static [FontId] {
     FONT_STACK.get().map(Vec::as_slice).unwrap_or(&[])
+}
+
+pub fn create_image_from_pixbuf(canvas: &mut Canvas<OpenGl>, image: &Pixbuf) -> Result<ImageId> {
+    let format = if image.has_alpha() {
+        PixelFormat::Rgba8
+    } else {
+        PixelFormat::Rgb8
+    };
+
+    let image_id = canvas.create_image_empty(
+        image.width() as usize,
+        image.height() as usize,
+        format,
+        ImageFlags::empty(),
+    )?;
+
+    // extract values
+    let width = image.width() as usize;
+    let stride = image.rowstride() as usize; // stride is in bytes per row
+    let height = image.height() as usize;
+    let bytes_per_pixel = if image.has_alpha() { 4 } else { 3 }; // pixbuf supports rgb or rgba
+
+    unsafe {
+        let src_buffer = image.pixels();
+
+        let row_length = width * bytes_per_pixel;
+        let mut dst_buffer = if row_length == stride {
+            // stride == row_length, there are no additional bytes after the end of each row
+            src_buffer.to_vec()
+        } else {
+            // stride != row_length, there are additional bytes after the end of each row that
+            // need to be truncated. We copy row by row..
+            let mut dst_buffer = Vec::<u8>::with_capacity(width * height * bytes_per_pixel);
+
+            for row in 0..height {
+                let src_offset = row * stride;
+                dst_buffer.extend_from_slice(&src_buffer[src_offset..src_offset + row_length]);
+            }
+            dst_buffer
+        };
+
+        // in almost all cases, that should be a no-op. Buf we might have additional elements after the
+        // end of the buffer, e.g. after width * height * bytes_per_pixel
+        dst_buffer.truncate(width * height * bytes_per_pixel);
+
+        if image.has_alpha() {
+            let img = Img::new_stride(
+                dst_buffer.align_to::<RGBA<u8>>().1.to_vec(),
+                width,
+                height,
+                width,
+            );
+
+            canvas.update_image(image_id, ImageSource::Rgba(img.as_ref()), 0, 0)?;
+        } else {
+            let img = Img::new_stride(
+                dst_buffer.align_to::<RGB<u8>>().1.to_owned(),
+                width,
+                height,
+                width,
+            );
+
+            canvas.update_image(image_id, ImageSource::Rgb(img.as_ref()), 0, 0)?;
+        }
+    }
+
+    Ok(image_id)
 }
 
 glib::wrapper! {
